cogs/owner.py
# Discord.py is smoooooooooooooosh!!!!!
import discord
from discord.ext import commands
import asyncio

import pymongo
import os

class Owner(commands.Cog):
    def __init__(self, bot):
        self.bot = bot


def setup(bot):
    bot.add_cog(Owner(bot))