requirements.txt
discord.py[voice]
flask
flask-discord
ImageHash
pya3rt
pillow
arrow
pymongo
colorama